backend/requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-community
neo4j
requests
streamlit
python-dotenv
pypdf
langchain-experimental
sentence-transformers
